requirements.txt
python-telegram-bot==22.2
python-dotenv==1.0.0
